SYNC-1540 Move
