+++ modules/util/sync-engine/src/main/java/com/liferay/sync/engine/documentlibrary/util/FileEventUtil.java
@@ -37,7 +37,7 @@ import com.liferay.sync.engine.model.SyncSite;
 import com.liferay.sync.engine.service.SyncFileService;
 import com.liferay.sync.engine.util.FileUtil;
 import com.liferay.sync.engine.util.PropsValues;
-import com.liferay.sync.engine.util.ReleaseInfo;
+import com.liferay.sync.engine.util.ServerInfo;
 
 import java.io.IOException;
 
@@ -274,7 +274,7 @@ public class FileEventUtil {
 
 		parameters.put("repositoryId", repositoryId);
 
-		if (ReleaseInfo.isServerCompatible(syncAccountId, 5)) {
+		if (ServerInfo.isServerCompatible(syncAccountId, 5)) {
 			parameters.put("retrieveFromCache", retrieveFromCache);
 		}
