+++ modules/util/sync-engine/src/main/java/com/liferay/sync/engine/util/ReleaseInfo.java
@@ -14,15 +14,6 @@
 
 package com.liferay.sync.engine.util;
 
-import com.liferay.sync.engine.documentlibrary.model.SyncContext;
-import com.liferay.sync.engine.model.SyncAccount;
-import com.liferay.sync.engine.service.SyncAccountService;
-
-import java.util.HashMap;
-import java.util.Map;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
 /**
  * @author Shinn Lok
  */
@@ -40,50 +31,6 @@ public class ReleaseInfo {
 		return _VERSION;
 	}
 
-	public static boolean isServerCompatible(
-		long syncAccountId, int minimumVersion) {
-
-		Boolean serverCompatible = _serverCompatibilityMap.get(
-			syncAccountId + "#" + minimumVersion);
-
-		if (serverCompatible != null) {
-			return serverCompatible;
-		}
-
-		SyncAccount syncAccount = SyncAccountService.fetchSyncAccount(
-			syncAccountId);
-
-		serverCompatible = isServerCompatible(
-			syncAccount.getPluginVersion(), minimumVersion);
-
-		_serverCompatibilityMap.put(
-			syncAccountId + "#" + minimumVersion, serverCompatible);
-
-		return serverCompatible;
-	}
-
-	public static boolean isServerCompatible(SyncContext syncContext) {
-		return isServerCompatible(syncContext.getPluginVersion(), 4);
-	}
-
-	protected static boolean isServerCompatible(
-		String pluginVersion, int minimumVersion) {
-
-		Matcher matcher = _pattern.matcher(pluginVersion);
-
-		if (!matcher.find()) {
-			return false;
-		}
-
-		if (pluginVersion.startsWith("6.2") &&
-			(Integer.parseInt(matcher.group(1)) < minimumVersion)) {
-
-			return false;
-		}
-
-		return true;
-	}
-
 	private static final String _BUILD = "3013";
 
 	private static final int _FEATURE_SET = 1;
@@ -91,10 +38,5 @@ public class ReleaseInfo {
 	private static final int _BUILD_NUMBER = Integer.parseInt(_BUILD);
 
 	private static final String _VERSION = "3.0.13";
-
-	private static final Pattern _pattern = Pattern.compile(
-		"(?:[0-9]+\\.){3}([0-9]+)");
-	private static final Map<String, Boolean> _serverCompatibilityMap =
-		new HashMap<>();
 
 }
